Refactor message extraction logic by introducing SequenceNodeProcessor to streamline node processing and enhance clarity
